requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
billiard==4.1.0
boto3==1.26.149
botocore==1.29.149
celery==5.3.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
daphne==4.0.0
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.0
msgpack==1.0.5
Pillow==9.5.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
